cyberrisk-platform/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
paho-mqtt==2.0.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
langgraph==0.1.1
langchain==0.2.1
langchain-community==0.2.1
httpx==0.27.0
python-dotenv==1.0.1
websockets==12.0
aiofiles==23.2.1
